+++ integration_test/lib/vnspec/api/vnctl.rb
@@ -6,7 +6,7 @@ module Vnspec
 
         args = build_args(method, url, params)
         address = config[:webapi][:host]
-        command = "cd #{config[:vnet_path]}/vnctl && bundle exec bin/vnctl #{args.join(" ")}"
+        command = "vnctl #{args.join(" ")}"
         logger.debug command
         logger.debug "params:"
         logger.debug old_params
@@ -16,21 +16,31 @@ module Vnspec
           raise "Request failed: #{raw_response[:stderr]}"
         end
 
-        YAML.load(raw_response[:stdout]).tap do |response|
-          logger.debug "response:"
-          logger.debug response
+        begin
+          response = YAML.load(raw_response[:stdout])
+        rescue
+          logger.error(
+            "An error occurred while trying to parse the following WebAPI reply as Yaml.\n%s" %
+            raw_response[:stdout])
 
-          if response.is_a?(Hash)
-            response = response.symbolize_keys
-            if response[:error]
-              raise "Request failed: #{response[:error]} #{response[:code]} #{response[:message]}"
-            end
-          elsif response.is_a?(String)
-            raise "Request failed: #{response}"
-          end
-
-          return yield(response) if block_given?
+          raise
         end
+
+        logger.debug "response:"
+        logger.debug response
+
+        if response.is_a?(Hash)
+          response = response.symbolize_keys
+          if response[:error]
+            raise "Request failed: #{response[:error]} #{response[:code]} #{response[:message]}"
+          end
+        elsif response.is_a?(String)
+          raise "Request failed: #{response}"
+        end
+
+        return yield(response) if block_given?
+
+        response
       end
 
       private
@@ -38,32 +48,36 @@ module Vnspec
         values = url.split("/").compact
         # datapath networks show dp-3 nw-public
         args = [values[0]]
-        args <<
-          case url
-          when %r(^datapaths/[^/]+/networks)
-            :networks
-          when %r(^datapaths/[^/]+/route_links)
-            :route_links
-          when %r(^interfaces/[^/]+/security_groups)
-            :security_groups
-          when %r(^dns_services/[^/]+/dns_records)
-            :dns_records
-          end
-        args += [convert_method(method), values[1], values[3]].compact
+
+        # Convert relation WebAPI calls to vnctl arguments.
+        # For example 'POST datapaths/dp-xxxx/networks/nw-yyyy'
+        # becomes: 'vnctl datapaths networks add dp-xxxx nw-yyyy'
+        relation_capture = url.match(%r([^/]+/[^/]+/([^/]+)))
+        if relation_capture
+          relation_name = relation_capture.captures.first
+
+          # Currently 'PUT interfaces/:uuid/rename' is the only API route that
+          # does not follow the REST standard. This is a quick hack to work
+          # around it.
+          args << relation_name unless relation_name == 'rename'
+        end
+
+        args += [convert_method(method, url), values[1], values[3]].compact
         params.keys.each do |key|
           args << %Q(--#{key} "#{params[key]}")
         end
+
         args
       end
 
-      def convert_method(method)
+      def convert_method(method, url)
         case method
         when :get
           :show
         when :post
           :add
         when :put
-          :modify
+          url =~ %r([^/]+/rename$) ? :rename : :modify
         when :delete
           :del
         else