+++ integration_test/vagrant/site-cookbooks/vnet/recipes/default.rb
@@ -1,8 +1,0 @@
-#
-# Cookbook Name:: vnet
-# Recipe:: default
-#
-# Copyright 2014, YOUR_COMPANY_NAME
-#
-# All rights reserved - Do Not Redistribute
-#