+++ vnet/lib/plugins/vdc_vnet_plugin.rb
@@ -114,7 +114,7 @@ module Vnet::Plugins
 
           if datapath_network.nil?
             Vnet::NodeApi::DatapathNetwork.create(datapath_network_params.merge({
-              :broadcast_mac_address => mac_generate(MAC_ADDRESS_PREFIX_DATAPATH_NETWORK)
+              :mac_address => mac_generate(MAC_ADDRESS_PREFIX_DATAPATH_NETWORK)
             }))
           end
         end
@@ -131,7 +131,7 @@ module Vnet::Plugins
 
       params = {
         :interface_id => simulated_interface.id,
-        :type => vnet_params[:name]
+        :mode => vnet_params[:name]
       }
       ns = Vnet::NodeApi::NetworkService.find(params)
       if ns.nil?