+++ vnet/lib/vnet/constants/filter.rb
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+module Vnet::Constants::Filter
+  MODE_STATIC = 'static'
+
+  MODES = [ MODE_STATIC ]
+end