+++ vnet/lib/vnet/endpoints/1.0/translations.rb
@@ -44,6 +44,8 @@ Vnet::Endpoints::V10::VnetAPI.namespace '/translations' do
     param :egress_ipv4_address, :String, transform: PARSE_IPV4, required: true
     param :ingress_port_number, :Integer, in: 1..65536
     param :egress_port_number, :Integer, in: 1..65536
+    param_uuid M::Network, :ingress_network_uuid
+    param_uuid M::Network, :egress_network_uuid
   end
 
   static_address_shared_params
@@ -68,7 +70,14 @@ Vnet::Endpoints::V10::VnetAPI.namespace '/translations' do
       egress_port_number: params["egress_port_number"]
     )
 
-    respond_with(R::TranslationStaticAddress.generate(tsa))
+    r = R::TranslationStaticAddress.generate(tsa)
+
+    if params['ingress_network_uuid'] && params['egress_network_uuid']
+      r[:ingress_network_uuid] = params['ingress_network_uuid']
+      r[:egress_network_uuid] = params['egress_network_uuid']
+    end
+
+    respond_with(r)
   end
 
   static_address_shared_params