+++ vnet/lib/vnet/model_wrappers/filter.rb
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+require 'ipaddress'
+
+module Vnet::ModelWrappers
+
+  class Filter < Base
+  end
+
+  class FilterStatic < Base
+    #include Helpers::IPv4
+  end
+
+end