+++ vnet/lib/vnet/node_api/datapath.rb
@@ -18,6 +18,7 @@ module Vnet::NodeApi
 
         # 0001_origin
         ActiveInterface.dispatch_deleted_where(default_filter, model.deleted_at)
+        ActiveNetwork.dispatch_deleted_where(default_filter, model.deleted_at)
         # datapath_network: ignore, handled by main event
         # datapath_route_link: ignore, handled by main event
         InterfacePort.dispatch_deleted_where(default_filter, model.deleted_at)